Move some define
