+++ src/gameobjects/units/cUnit.h
@@ -15,6 +15,7 @@
 #include "gameobjects/structures/cAbstractStructure.h"
 #include "player/brains/missions/cPlayerBrainMission.h"
 #include "structs.h"
+#include "gameobjects/units/cUnitInfo.h"
 #include "utils/cRectangle.h"
 
 #include <memory>
