+++ src/player/cPlayer.h
@@ -28,6 +28,7 @@
 #include "player/brains/cPlayerBrain.h"
 #include "player/cPlayerDifficultySettings.h"
 #include "sidebar/cSideBar.h"
+#include "gameobjects/units/cUnitInfo.h"
 
 #include <set>
 #include <string>
