+++ src/sidebar/cBuildingListItem.h
@@ -14,6 +14,7 @@
 class cBuildingList;
 struct s_SpecialInfo;
 struct s_UpgradeInfo;
+struct s_UnitInfo;
 
 class cBuildingListItem {
